issue37: contextmenu -> widget. draft
