+++ src/lib/leaflet.control.export/export-formats.js
@@ -3,11 +3,28 @@ import {RmapsWriter} from './rmaps-encoder';
 
 const formats = [];
 
-function addFormat(encoder, name, hint, fileExtension) {
-    formats.push({encoder: encoder, name: name, hint: hint, fileExtension: fileExtension});
+function createBaseExportOption(params) {
+    const {format, layer, layerName, zoom, metersPerPixel, tilesCount, fileSizeMb} = params;
+    let resolutionString = metersPerPixel.toFixed(2);
+    let sizeString = fileSizeMb.toFixed(fileSizeMb > 1 ? 0 : 1);
+
+    params.label = `Zoom ${zoom} (${resolutionString} m/pixel) &mdash; ${tilesCount} tiles (~${sizeString} Mb)`;
+    params.itemClass = '';
+    params.tooltip = '';
+    return params;
 }
 
-addFormat(JnxWriter, "JNX", "Garmin", "jnx");
-addFormat(RmapsWriter, "RMaps", "OsmAnd", "sqlitedb");
+function addFormat(encoder, name, hint, fileExtension, exportOptionFactory) {
+    formats.push({encoder, name, hint, fileExtension, exportOptionFactory});
+}
+
+addFormat(JnxWriter, "JNX", "Garmin", "jnx", function(params) {
+    const option = createBaseExportOption(params);
+    option.itemClass = option.tilesCount > 50000 ? 'export-menu-warning' : '';
+    option.tooltip = option.tilesCount > 50000 ? '> 50000 tiles' : '';
+    return option;
+});
+
+addFormat(RmapsWriter, "RMaps", "OsmAnd", "sqlitedb", createBaseExportOption);
 
 export {formats};
