--- src/lib/leaflet.control.export/index.js
@@ -1,209 +1,0 @@
-import L from 'leaflet';
-import ko from 'vendored/knockout';
-import './style.css';
-import 'lib/leaflet.control.commons';
-import {RectangleSelect} from './selector';
-import Contextmenu from 'lib/contextmenu';
-import {exportFromLayer, minZoom} from './export-maker';
-import {formats} from './export-formats';
-import localStorage from 'lib/safe-localstorage';
-
-import {saveAs} from 'vendored/github.com/eligrey/FileSaver';
-import {notify} from 'lib/notifications';
-import logging from 'lib/logging';
-
-const ATTR_EXPORT_FORMAT = "exportFormat";
-
-function saveFormatNameToLocalStorage(formatName) {
-    localStorage.setItem(ATTR_EXPORT_FORMAT, formatName);
-}
-
-function getFormatNameFromLocalStorage() {
-    return localStorage.getItem(ATTR_EXPORT_FORMAT);
-}
-
-L.Control.export = L.Control.extend({
-        includes: L.Mixin.Events,
-
-        initialize: function(layersControl, options) {
-            L.Control.prototype.initialize.call(this, options);
-            this._layersControl = layersControl;
-            this.exportInProgress = ko.observable(false);
-            this.downloadProgressRange = ko.observable(1);
-            this.downloadProgressDone = ko.observable(0);
-            this.contextMenu = new Contextmenu(() => this.makeMenuItems());
-        },
-
-    getExportFormat: function() {
-        if (this.exportFormat === undefined) {
-            let fromStorage = getFormatNameFromLocalStorage();
-            this.exportFormat = formats.find( (f) => (f.name == fromStorage) ) || formats[0];
-        }
-        return this.exportFormat;
-    },
-
-    getExportFormatName: function() {
-        return this.getExportFormat().name;
-    },
-
-    setExportFormatByName: function(formatName) {
-        this.exportFormat = formats.find( (f) => (f.name == formatName ) );
-            if ( this.exportFormat === undefined ) {
-                this.exportFormat = this.getExportFormat();
-                this.fire('exportformatchange');
-            }
-            saveFormatNameToLocalStorage(this.getExportFormatName());
-        },
-
-        getExportLayer: function() {
-            let selectedLayer = {};
-            for (let layerRec of this._layersControl._layers) {
-                let layer = layerRec.layer;
-                if (this._map.hasLayer(layer) && layer.options && layer.options.export) {
-                    selectedLayer = {
-                        layer,
-                        layerName: layerRec.name
-                    }
-                }
-            }
-            return selectedLayer;
-        },
-
-        estimateTilesCount: function(maxZoom) {
-            let tilesCount = 0;
-            const bounds = this._selector.getBounds();
-            for (let zoom=minZoom(maxZoom); zoom <= maxZoom; zoom++) {
-                const topLeftTile = this._map.project(bounds.getNorthWest(), zoom).divideBy(256).floor();
-                const bottomRightTile = this._map.project(bounds.getSouthEast(), zoom).divideBy(256).ceil();
-                tilesCount += Math.ceil((bottomRightTile.x - topLeftTile.x) * (bottomRightTile.y - topLeftTile.y));
-            }
-            return tilesCount;
-        },
-
-        makeMenuItems: function() {
-            const {layer, layerName} = this.getExportLayer();
-            if (!layer) {
-                return [{text: 'No supported layers'}];
-            }
-            const maxLevel = layer.options.maxNativeZoom || layer.options.maxZoom || 18;
-            const minLevel = Math.max(0, maxLevel - 6);
-
-            const equatorLength = 40075016;
-            const lat = this._selector.getBounds().getCenter().lat;
-            let metersPerPixel = equatorLength / Math.pow(2, maxLevel) / 256 * Math.cos(lat / 180 * Math.PI);
-            let changeFormat = (formatName) => {
-                this.setExportFormatByName(formatName);
-                this.fire('exportformatchange');
-            }
-            const items = [
-                {text: "Output format", header: true},
-                {
-                    selectmenu: true,
-                    values: formats.map(function(x) {return {value: x.name, label: `${x.name} (${x.hint})`}}),
-                    callback: changeFormat, 
-                    defaultValue: this.getExportFormatName()
-                },
-                {text: layerName, header: true}
-            ];
-            for (let zoom = maxLevel; zoom >= minLevel; zoom -= 1) {
-                let tilesCount = this.estimateTilesCount(zoom);
-                let fileSizeMb = tilesCount * 0.02;
-                let itemClass = tilesCount > 50000 ? 'export-menu-warning' : '';
-                let resolutionString = metersPerPixel.toFixed(2);
-                let sizeString = fileSizeMb.toFixed(fileSizeMb > 1 ? 0 : 1);
-                let item = {
-                    text: `<span class="${itemClass}">Zoom ${zoom} (${resolutionString} m/pixel) &mdash; ${tilesCount} tiles (~${sizeString} Mb)</span>`,
-                    callback: () => this.startExport(this.getExportFormat(), layer, layerName, zoom),
-                    disabled: this.exportInProgress()
-                };
-                items.push(item);
-                metersPerPixel *= 2;
-            }
-            return items;
-        },
-
-        notifyProgress: function(value, maxValue) {
-            this.downloadProgressDone(this.downloadProgressDone() + value);
-            this.downloadProgressRange(maxValue);
-        },
-
-        startExport: function(format, layer, layerName, zoom) {
-            logging.captureBreadcrumbWithUrl({message: 'start export'});
-            this.exportInProgress(true);
-            this.downloadProgressDone(0);
-
-            const bounds = this._selector.getBounds();
-            const sanitizedLayerName = layerName.toLowerCase().replace(/[ ()]+/, '_');
-            const fileName = `nakarte.me_${sanitizedLayerName}_z${zoom}.${format.fileExtension}`;
-            const eventId = logging.randId();
-            logging.logEvent('export start', {eventId, layerName, zoom, bounds});
-            exportFromLayer(format, layer, layerName, zoom, bounds, this.notifyProgress.bind(this))
-                .then((fileData) => {
-                    saveAs(fileData, fileName, true);
-                    logging.logEvent('export end', {eventId, success: true});
-                })
-                .catch((e) => {
-                        logging.captureException(e);
-                        logging.logEvent('export end', {eventId, success: false, error: e.stack});
-                        notify(`Export failed: ${e.message}`);
-                    }
-                )
-                .then(() => this.exportInProgress(false));
-        },
-
-        onAdd: function(map) {
-            this._map = map;
-            const container = this._container = L.DomUtil.create('div', 'leaflet-control leaflet-control-export');
-            container.innerHTML = `
-                <a class="button" data-bind="visible: !exportInProgress(), click: onButtonClicked"
-                 title="Export tiles for GPS receivers">Export</a>
-                <div data-bind="
-                    component:{
-                        name: 'progress-indicator',
-                        params: {progressRange: downloadProgressRange, progressDone: downloadProgressDone}
-                    },
-                    visible: exportInProgress()"></div>`;
-            ko.applyBindings(this, container);
-            this._stopContainerEvents();
-            return container;
-        },
-
-        removeSelector: function() {
-            if (this._selector) {
-                this._map.removeLayer(this._selector);
-                this._selector = null;
-                this.fire('selectionchange');
-            }
-        },
-
-        addSelector: function(bounds) {
-            if (!bounds) {
-                bounds = this._map.getBounds().pad(-0.25);
-            }
-            this._selector = new RectangleSelect(bounds)
-                .addTo(this._map)
-                .on('change', () => this.fire('selectionchange'))
-                .on('click contextmenu', (e) => {
-                    L.DomEvent.stop(e);
-                    this.contextMenu.show(e)
-                });
-            this.fire('selectionchange');
-        },
-
-
-        onButtonClicked: function() {
-            if (this._selector) {
-                if (this._selector.getBounds().intersects(this._map.getBounds().pad(-0.05))) {
-                    this.removeSelector();
-                } else {
-                    this.removeSelector();
-                    this.addSelector();
-                }
-
-            } else {
-                this.addSelector();
-            }
-        },
-    }
-);
-
